Mongo Aggregation Better

+++ [Theme] NoSQL, Redis, MongoDB/Mongo Aggregation, $group, $project, $min, $max, $avg, $size,$sum/src/main/java/mongo/Store.java
@@ -14,7 +14,7 @@ public class Store {
     private double avgPrice;
     private int biggestPrice;
     private int lowestPrice;
-    private int numberOfProducts;
+    private Long numberOfProducts;
     private Long numberOfGreaterThanProducts;
 
     public Store(String storeName) {
